flask_app/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
groq==0.18.0
python-dotenv==0.21.0